Finished integration tutorial
