+++ Sources/MountebankSwift/Documentation.docc/Resources/integration/FeedUITests_12_previous.swift
@@ -1,0 +1,26 @@
+let listsImposter = Imposter(
+    name: "lists",
+    stubs: [
+        Stub(
+            responses: [
+                Is(statusCode: 500),
+                Is(statusCode: 200, body: ["podcasts": []]),
+            ]
+            predicate: .equals(Request(method: .get, path: "/trending.json"))
+        ),
+    ]
+)
+
+guard let listsPort = try await mountebank.postImposter(imposter: listsImposter).port else {
+    XCTFail("Port should have been set by now")
+    return
+}
+
+let app = XCUIApplication()
+app.launchEnvironment["listApiHost"] = "http://localhost:\(listsPort)"
+app.launch()
+
+XCTAssertTrue(app.staticTexts["Error: 500"].exists)
+
+app.buttons["Retry"].tap()
+XCTAssertTrue(app.staticTexts["No podcasts found"].exists)
