+++ Sources/MountebankSwift/Documentation.docc/Resources/integration/FeedUITests_13_previous.swift
@@ -1,0 +1,8 @@
+let app = XCUIApplication()
+app.launchEnvironment["listApiHost"] = "http://localhost:\(listsPort)"
+app.launch()
+
+XCTAssertTrue(app.staticTexts["Error: 500"].exists)
+
+app.buttons["Retry"].tap()
+XCTAssertTrue(app.staticTexts["No podcasts found"].exists)
