+++ Sources/MountebankSwift/Documentation.docc/Resources/integration/FeedUITests_16_previous.swift
@@ -1,0 +1,12 @@
+let app = XCUIApplication()
+app.launchEnvironment["listApiHost"] = "http://localhost:\(listsPort)"
+app.launchEnvironment["imagesApiHost"] = "http://localhost:\(imagesPort)"
+app.launch()
+
+XCTAssertTrue(app.staticTexts["Error: 500"].exists)
+
+app.buttons["Retry"].tap()
+XCTAssertTrue(app.staticTexts["No podcasts found"].exists)
+
+app.buttons["Retry"].tap()
+XCTAssertTrue(app.staticTexts["Serial"].exists)
