+++ Sources/MountebankSwift/Documentation.docc/Resources/integration/sample_app_host_images.swift
@@ -1,16 +1,18 @@
 enum Host {
-    // URL that will only be not nil during UI testing
     static var uiTestingListApiHost: URL? {
         ProcessInfo.processInfo.environment["listApiHost"].flatMap(URL.init(string:))
     }
+    static var uiTestingImagesApiHost: URL? {
+        ProcessInfo.processInfo.environment["imagesApiHost"].flatMap(URL.init(string:))
+    }
 
     static var lists: URL {
-        // Use UI testing URL if it is set or use production URL otherwise
         uiTestingListApiHost ??
         URL(string: "https://lists.pocketcasts.com")!
     }
 
     static var images: String {
+        uiTestingImagesApiHost ??
         URL(string: "https://static.pocketcasts.com")!
     }
 }
