+++ Sources/MountebankSwift/Documentation.docc/Resources/integration/sample_app_host_images_initial.swift
@@ -1,11 +1,9 @@
 enum Host {
-    // URL that will only be not nil during UI testing
     static var uiTestingListApiHost: URL? {
         ProcessInfo.processInfo.environment["listApiHost"].flatMap(URL.init(string:))
     }
 
     static var lists: URL {
-        // Use UI testing URL if it is set or use production URL otherwise
         uiTestingListApiHost ??
         URL(string: "https://lists.pocketcasts.com")!
     }
